Fixing Deletion bug
