+++ src/App.tsx
@@ -3,8 +3,7 @@ import React, {useRef, useState} from "react";
 import "./App.css";
 import TableCourseInput from "./components/tables/TableCourseInput";
 import "react-dropdown/style.css";
-import Drop from "./components/dropdown/Drop";
-import { getScale } from "./components/dropdown/GpaDrop";
+import {Drop, getScale} from "./components/dropdown/Drop";
 import WebHeader from "./components/WebHeader";
 import {
     Button,
@@ -180,6 +179,7 @@ const App = () => {
                                     </Grow>
                                 )}
                             </Popper>
+                            
                             {/* Button to select scale */}
                             <Button
                                 size="large"
