+++ src/components/dropdown/Drop.tsx
@@ -1,18 +1,45 @@
 import React, { useEffect, useRef, useState } from "react";
-import { GpaDrop } from "./GpaDrop";
 import Select, { SelectChangeEvent } from "@mui/material/Select";
 import { FormControl, InputLabel, MenuItem } from "@mui/material";
 import InfoTooltip from "./InfoTooltip";
 import { CSSTransition } from "react-transition-group";
 
+let getScale: string;
+
 const Drop = () => {
     //Setting Country
     const [country, setCountry] = useState("");
 
-    const handleChange = (e: SelectChangeEvent) => {
+    const countryRef = useRef('');
+
+    const handleCountryChange = (e: SelectChangeEvent) => {
         setCountry(e.target.value)
+        setScale('')
+        countryRef.current = e.target.value
     };
 
+    const [scale, setScale] = useState('');
+
+    // Updating export value
+    const handleScaleChange = (e: SelectChangeEvent) => {
+        setScale(e.target.value)
+        getScale = e.target.value
+    }
+
+    //Hides India scale when Country isn't India
+    const indiaHidden = () => {
+        if(country !== 'India')
+            return true
+        return false
+    }
+
+    //Hides China scale when Country isn't China
+    const chinaHidden = () => {
+        if(country !== 'China')
+            return true
+        return false
+    }
+    
     return (
         <div id="gpa-dropdown">
             <div className="alignLabel">
@@ -22,7 +49,7 @@ const Drop = () => {
                     <Select
                         id="dropdown-selector"
                         value={country}
-                        onChange={handleChange}
+                        onChange={handleCountryChange}
                         label={"Country"}
                     >
                         {/* Selections of Countries from Dropdowns */}
@@ -32,22 +59,44 @@ const Drop = () => {
                 </FormControl>
             </div>
                 <CSSTransition
-                    in={country != ""}
+                    in={countryRef.current != ""}
                     timeout={300}
                     classNames="fade"
                     unmountOnExit
                 >
                     <div className="alignLabel" style={{width: '100vh', display: 'flex', flexDirection: 'row'}}>
                         <label htmlFor="scale-dropdown">Grading Scale :&emsp;</label>
-                        <GpaDrop id="scale-dropdown" country={country}/>
 
-                        <InfoTooltip country={country}/>
 
-                        {/* Passing the country value another way than this way */}
+
+                        <FormControl size="small" sx={{ width: 'auto', minWidth: 146}}>
+                <InputLabel>Grading Scale</InputLabel>
+                <Select
+                    defaultValue={''}
+                    id="dropdown-selector"
+                    value={scale}
+                    onChange={handleScaleChange}
+                    label={'Grading Scale'}
+                    autoWidth={true}
+                >
+                    {/* India Scales */}
+                    <MenuItem hidden={indiaHidden()} value={"Most Common Scale"}>Most Common Scale</MenuItem>  
+                    <MenuItem hidden={indiaHidden()} value={"Letter Grade Scale"}>Letter Grade Scale</MenuItem>  
+                    <MenuItem hidden={indiaHidden()} value={"Choice Based System"}>Choice Based Credit Scale</MenuItem>  
+                    <MenuItem hidden={indiaHidden()} value={"IIT Scale"}> IIT Scale</MenuItem> 
+                    {/* China Scales */}
+                    <MenuItem hidden={chinaHidden()} value={"5 Point Scale"}>5 Point Scale</MenuItem>  
+                    <MenuItem hidden={chinaHidden()} value={"4 Point Scale"}>4 Point Scale</MenuItem>  
+
+                </Select>
+            </FormControl>
+
+                        <InfoTooltip country={countryRef.current}/>
+
                     </div>
                 </CSSTransition>
         </div>
     );
 };
 
-export default Drop;
+export{Drop, getScale};
